+++ source/extensions/omni.isaac.lab/omni/isaac/lab/sensors/ray_caster/ray_caster_camera_cfg.py
@@ -5,14 +5,14 @@
 
 """Configuration for the ray-cast camera sensor."""
 
-from typing import Literal
 from dataclasses import MISSING
+from typing import Literal
 
 from omni.isaac.lab.utils import configclass
 
+from .patterns import PinholeCameraPatternCfg
 from .ray_caster_camera import RayCasterCamera
 from .ray_caster_cfg import RayCasterCfg
-from .patterns import PinholeCameraPatternCfg
 
 
 @configclass